split pandas/tests/base/test_ops.py
